refactor(web): 토끼 코드 리뷰반영
